+++ src/aigora/curriculum_graph/application/loading/graph_loader.py
@@ -8,9 +8,7 @@ from aigora.curriculum_graph.application.assembling.graph_assembler import Graph
 from aigora.curriculum_graph.application.parsing.graph_parser import GraphParser
 from aigora.curriculum_graph.application.mapping.graph_mapper import GraphMapper
 from aigora.curriculum_graph.application.validation.graph_validator import GraphValidator
-
-from .loader_errors import GraphLoaderError
-
+from aigora.curriculum_graph.application.loading.loader_errors import GraphLoaderError
 
 class GraphLoader:
     """Loads a CurriculumGraph from a file-based source.
@@ -76,7 +74,7 @@ class GraphLoader:
     def _map_version(self, payload: dict[str, Any]) -> str | None:
         version = payload.get("version")
         if version is not None and not isinstance(version, str):
-            from .mapping.mapper_errors import InvalidGraphPayloadError
+            from ..mapping.mapper_errors import InvalidGraphPayloadError
             raise InvalidGraphPayloadError(
                 "Graph payload field 'version' must be a string."
             )
